Unity Application For Raspberry Pi 4
Almost there. I have optimized the graphics and configuration settings for raspberry pi 4. It is able to output 100 image slices without animation changes. The reason I said almost is because the supported Graphics API [OpenGL-Core] that I can generate on Windows is not the same as Raspbian OS [OpenGLES3]. Can render but introduces a major artifacts. 3 Options as of now

Option 1: Build the unity application on the Raspberry PI 4 and hope the supported graphics API is there

Option 2: Go the dumb route and use WebGL or Android
** I have heard that it is not a good idea to use WebGL as it is browser based on Chromium. So there's overhead with running a google chrome tab and I heard the image manipulation is difficult and unreliable.
** I dont want to use Android because it requires LineageOS and runs on that. I dont know how we will do the LED driver code if that is the case

Option 3: Scrap unity as a image generator and use unity as a frame-by-frame animation generator to perform animation states on the LED driver code by calling specific subdirectories containing the animation state. In this case, unity will simply only be used for frame generation and the main algorithm of the hologram will be contained in the LED driver code. This will require a state machine to be defined to determine which animation state and which one is loop forever until told overwise

Also I am just doing this for my own sanity (and bongo cat)